requirements.txt
tk
customtkinter
pyaudio
SpeechRecognition
pydub
pyttsx3
pillow
pygame